requirements.txt
discord.py~=1.6.0
python-dotenv~=0.15.0
pampy~=0.3.0
JSON-log-formatter~=0.3.0